praying it works

+++ backend/src/main/java/com/ilpalazzo/errors/OrderNotFoundException.java
@@ -2,7 +2,7 @@ package com.ilpalazzo.errors;
 
 public class OrderNotFoundException extends RuntimeException {
 
-    public OrderNotFoundException(Integer id) {
+    public OrderNotFoundException(Long id) {
         super("Order not found with id: " + id);
     }
 }
